Added additional multiclass capabilities
Can now enter multiple numbers of workcrews in the crew argument for
find_sequence.py. With the input -d 1 2 3 4 the code will perform all
calculations for 1 workcrew and then postprocess for 2, 3, and 4 and
output a list of the objective functions.
Added multiclass implementations of the greedy and simulated annealing
heuristics.
Added eval_working_sequence to find_sequence.py which allows use of
memory (unlike eval_sequence in sequence_utils.py) and speeds up the
multiclass implementation of simulated annealing.
Beam search does not yet have a multiclass implementation, therefore
it should be run for 1 workcrew and postprocessed for multiple crews.
The arguments -d 1 2 3 4 --opt True will result in four seperate runs
of the brute force algorithm to find optimal solutions for each number
of workcrews.